requirements.txt
Django==5.1.2
django-model-utils==4.5.1

psycopg2-binary==2.9.9
